requirements.txt
python-vlc
pafy
flask
flask_cors
youtube_dl
